backend/requirements.txt
Django==4.2.11
djangorestframework==3.15.1
djoser==2.2.2
psycopg2-binary==2.9.9
Pillow==10.3.0
django-cleanup==8.1.0
django-filter==2.4.0
gunicorn==21.2.0
fpdf2==2.7.8
uharfbuzz==0.39.1
drf_extra_fields==3.7.0
python-dotenv
django-cors-headers==3.13.0
django-colorfield==0.11.0
webcolors==1.11.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
PyYAML==6.0
hashids